backend/requirements.txt
django==5.2.8
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
psycopg2-binary==2.9.11
python-decouple==3.8
gunicorn==23.0.0
django-cors-headers==4.6.0
